refactor: rename minio package to nodepool
