+++ cmd/serve/errors.go
@@ -1,0 +1,5 @@
+package serve
+
+var (
+	errNodesNotFound = errors.New("minio nodes not found")
+)
