+++ pkg/nodepool/nodeconfig.go
